CMS-949-bulk-tagger - adding process classes
