+++ src/Operations/TermManagerOperationItem.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace Drupal\dennis_term_manager;
+namespace Drupal\dennis_term_manager\Operations;
+
+use Drupal\dennis_term_manager\TermManagerItem;
+
 /**
  * TermManagerOperationItem
  */
