--- src/Progress/TermManagerExport.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Drupal\dennis_term_manager\Progress;
-
-
-class TermManagerExport {
-
-}
