+++ src/Progress/TermManagerProgressList.php
@@ -4,12 +4,33 @@
 namespace Drupal\dennis_term_manager\Progress;
 
 
+use Drupal\Core\Messenger\Messenger;
+use Drupal\Core\Config\ConfigFactoryInterface;
+
+
 /**
- * TermManagerProgressList
+ * Class TermManagerProgressList
+ *
+ * @package Drupal\dennis_term_manager\Progress
  */
 class TermManagerProgressList implements \Iterator, \Countable {
 
 
+  /**
+   * The config factory.
+   *
+   * @var \Drupal\Core\Config\ConfigFactoryInterface
+   */
+  protected $configFactory;
+
+  /**
+   * @var \Drupal\Core\Messenger\Messenger
+   */
+  protected $messenger;
+
+  /**
+   * @var TermManagerProgressItem
+   */
   protected $termManagerProgressItem;
 
   /**
@@ -29,38 +50,29 @@ class TermManagerProgressList implements \Iterator, \Countable {
    *
    * TermManagerProgressList constructor.
    *
+   * @param ConfigFactoryInterface $configFactory
+   * @param Messenger $messenger
    * @param TermManagerProgressItem $termManagerProgressItem
    */
-  public function __construct(TermManagerProgressItem $termManagerProgressItem) {
-
+  public function __construct(ConfigFactoryInterface $configFactory,
+                              Messenger $messenger,
+                              TermManagerProgressItem $termManagerProgressItem) {
+    $this->configFactory = $configFactory;
+    $this->messenger = $messenger;
     $this->termManagerProgressItem = $termManagerProgressItem;
 
     $this->position = 0;
 
-
-    //TODO - fix the variable thingy
-    // Load list of current processes.
-  //  $in_progress = variable_get('dennis_term_manager_in_progress', []);
-
-    $config = \Drupal::config('dennis_term_manager');
-    $in_progress = $config->get('in_progress');
-
-
-    //$in_progress = [];
-
-
-
-
+    $in_progress =  $this->configFactory->get('dennis_term_manager')->get('in_progress', []);
 
     foreach (array_keys($in_progress) as $fid) {
       try {
-
-
-        $progress_item = $this->termManagerProgressItem->init($fid);
+        $progress_item = $this->termManagerProgressItem;
+        $progress_item->init($fid);
         $this->progressList[] = $progress_item;
       }
       catch (\Exception $e) {
-        \Drupal::messenger()->addMessage($e->getMessage());
+        $this->messenger->addMessage($e->getMessage());
       }
     }
   }
